add shipping functionalities
If a user is logged in, the projects which ship to the country of origin appear first and are sorted in alphabetical order, then followed by the rest of the projects with a tag. Users will be alerted in ViewProject if the project does not ship to the country of origin.